refactor: 메서드 책임 이동
